backend/requirements.txt
fastapi>=0.128.0
uvicorn>=0.27.0
anthropic>=0.86.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
sympy>=1.12
numpy>=1.26.0
chromadb>=0.4.22
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
youtube-transcript-api>=0.6.0
bcrypt>=4.0.0
pydantic[email]>=2.0.0
